Improvements to context parsing:
- Centralise 'parse context'

Rather than having seperate places where global variables are set/get (`_root`, `_token_types`) we centralize this concept into the `parse_context` module. Methods in this module are used to initialise, amend and retrieve the current set of block/span tokens, and the dict of gathered link definitions, which are parsed between funtions.

- Set the intialised parse context on `__enter__` of a renderer.

Before this, the context was reset on exit and would require re-initiation to estabilish the required tokens.

- Rename footnote -> link_definition

This has been renamed since, these actually refer to
https://spec.commonmark.org/0.28/#link-reference-definitions,
rather than what would generally be considered a footnote:
https://www.markdownguide.org/extended-syntax/#footnotes

- Make core_tokens._code_matches thread-safe