Segregate parsing and initialisation of block tokens
This is a major change to the block parsing logic. Rather than spreading the parsing over the the `read` classmethod and the `__init__` method, we contain all the parsing to `read`. This means that tokens can be initialised independently to document parsing, and we can introduce things like parameter documentation and validation.

The main issue faced, was that span tokens can not be parsed until all of the link definitions have been stored. To achieve this, we introduce the `SpanContainer` which simply stores the inline text, until the full documentation has been parsed. Then, in `tokenize_main` we walk through the document and expand any of these instances that we find.

We also allow for the storage of the actual `LinkDefinition` token. But don't by default, so the renderers are not affected.

All tests still pass, with some minimal changes.